+++ admin/headlines-settings.php
@@ -10,10 +10,22 @@ function dwrafh_remove_category() {
 	// Finds all posts with the category headline. It shouldn't be too many as there are typically only 3 ever which are live and up to 15 which can be set.  -JMS
 
 	$args = array(
-		'post_type'              => 'post',
+		'post_type'              => array('post', 'dice_ideal_employer'),
 		'posts_per_page'         => -1,
 		'post_status'            => array( 'publish', 'future' ),
-		'category_name'          => 'headline'
+		'tax_query' => array(
+			'relation' => 'OR',
+			array(
+				'taxonomy' => 'ideal-employer-category',
+				'field'    => 'slug',
+				'terms'    => 'headline'
+			),
+			array(
+				'taxonomy' => 'category',
+				'field'    => 'slug',
+				'terms'    => 'headline'
+			)
+		)
 	);
 
 	// Gets the query response. -JMS
@@ -74,14 +86,26 @@ function dwrafh_display_dashboard_widget() {
 
 	// Set up some query parameters. -JMS
 	$args = array(
-		'post_type'              => 'post',
+		'post_type'              => array( 'post', 'dice_ideal_employer' ),
 		'orderby'                => 'menu_order',
 		'order'                  => 'ASC',
 		'no_found_rows'          => true,
 		'update_post_term_cache' => false,
 		'posts_per_page'         => 15,
 		'post_status'            => array( 'publish', 'future' ),
-		'category_name'          => 'headline'
+		'tax_query' => array(
+			'relation' => 'OR',
+			array(
+				'taxonomy' => 'ideal-employer-category',
+				'field'    => 'slug',
+				'terms'    => 'headline'
+			),
+			array(
+				'taxonomy' => 'category',
+				'field'    => 'slug',
+				'terms'    => 'headline'
+			)
+		)
 	);
 
 	// Get the query response. -JMS
@@ -290,9 +314,9 @@ function dwrafh_remove_headline_cat() {
 
 	$headlinePostId = $_POST['parentId'];
 	$terms = 'headline';
-	$taxonomy = 'category';
 
-	wp_remove_object_terms( $headlinePostId, $terms, $taxonomy );
+	wp_remove_object_terms( $headlinePostId, $terms, 'category' );
+	wp_remove_object_terms( $headlinePostId, $terms, 'ideal-employer-category' );
 	wp_send_json_success( 'Post with ID of '. $headlinePostId .' has been removed' );
 
 	return null;